Add a few other typo fixes
I used a VS code extension for this, but it has one problem: it's only "en" not "en-us"

I plan on going through the rest of this file with that extension later, but keep in mind: Things will be spelled as if I'm not from 'merica for some time.
Those will be needed to be pointed out.